requirements.txt
# Core LangChain & LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langsmith>=0.2.0

# Mistral AI
mistralai>=1.0.0
langchain-mistralai>=0.1.0

# Web Search (optional)
tavily-python>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data & ML
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.1.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Visualization & Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter Notebook
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipython>=8.12.0
ipywidgets>=8.0.0

# UI
streamlit>=1.31.0
rich>=13.7.0
click>=8.1.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
